Add "sharedWithGroups" to Group
Fixes https://github.com/gitlab4j/gitlab4j-api/issues/1053
