+++ src/test/resources/org/gitlab4j/api/project.json
@@ -73,13 +73,15 @@
       "group_id": 4,
       "group_name": "Twitter",
       "group_full_path": "twitter",
-      "group_access_level": 30
+      "group_access_level": 30,
+      "expires_at": "2023-11-19"
     },
     {
       "group_id": 3,
       "group_name": "Gitlab Org",
       "group_full_path": "gitlab-org",
-      "group_access_level": 10
+      "group_access_level": 10,
+      "expires_at": "2023-11-19"
     }
   ],
   "repository_storage": "default",
